meta_agent_system/requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
pandas>=1.3.0
matplotlib>=3.4.0
pydantic>=2.0.0
colorlog>=6.0.0